+++ docker_env/application/source/login/register_form.php
@@ -1,11 +1,67 @@
 <?php
 
-use PHPMailer\PHPMailer\PHPMailer;
-use PHPMailer\PHPMailer\SMTP;
+	session_start();
+	require('log.php');
 
-require('src/connect.php');
+	// Here we check a session is already ON
+	if (isset($_SESSION['connect']))
+	{
+		header('location: index_login.php');
+		exit();
+	}
 
+	
+	if (!empty($_POST['email']) && !empty($_POST['password']) && !empty($_POST['password_two']))
+	{
+		include('../src/connect.php');
+		// variables
+		$email 					= htmlspecialchars($_POST['email']);
+		$password 				= htmlspecialchars($_POST['password']);
+		$password_two  			= htmlspecialchars($_POST['password_two']);
+		 
+		// password = password_two ?
+		if ($password != $password_two)
+		{
+			header('location: register_form.php?error=1&message=Non identical password.');
+			exit();
+		}
+
+		// email adress valid ?
+		if(!filter_var($email, FILTER_VALIDATE_EMAIL))
+		{
+			header('location : register_form.php?error=1&message=Email adress invalid');
+			exit();
+		}
+
+		// email already exists ?
+		$req = $db->prepare("SELECT count(*) as numberEmail FROM user WHERE email = ?");
+		$req->execute(array($email));
+
+		while($email_verification = $req->fetch())
+		{
+			if ($email_verification['numberEmail'] != 0)
+			{
+				header('location: register_form.php?error=1&message=Email already used.');
+				exit();
+			}
+		}
+
+		// hash
+		$secret = sha1($email);
+
+		// encryption password 
+		$password = "aq1".sha1($password."123")."25";
+
+		// sending
+		$req = $db->prepare("INSERT INTO user(email, password, secret) VALUES (?,?,?)");
+		$req->execute(array($email,$password,$secret));
+		header('location: register_form.php?success=1');
+		exit();
+
+	}
 ?>
+
+
 
 <!DOCTYPE html>
 <head>
@@ -15,6 +71,7 @@ require('src/connect.php');
     <link href="./styles/styles.css" rel="stylesheet">
 </head>
 <body>
+
 
 
 <div class="container1">
@@ -37,19 +94,36 @@ require('src/connect.php');
 		c-9.495,3.001-13.041,1.755-13.428,1.221C10.147,81.847,10.132,77.877,16.556,69.375L16.556,69.375z"/>
 </g>
 <rect x="0.118" y="0.11" fill="none" width="115.712" height="106.615"/>
-</svg>
+    </svg>
     </div>
-        <br/>
-        <h1 class="txtHello">Lost in our Galaxy ?</h1>
-        <h2 class="txtHello">Yes, send me a new password please !</h2>
-    <form method="post">
-            <input type="email" class="Register1_loginForm" name="email" label="Register" id="Register1_loginForm" placeholder="type your email adress..."/>
-        <div class="mx-auto" style="width:100px";>
-            <button type="submit" name="forgot_password" class="btn btn-secondary">Click Here</button>	
-        </div>
-	
+
+
+        <h1 class="txtHelloRegister">You're a star,<br> join our Galaxy!</h1>
+
+
+<!-- HERE PHP  -->
+<?php if (isset($_GET['error']))
+			{
+				if (isset($_GET['message']))
+				{
+					echo'<div>'.htmlspecialchars($_GET['message']).'</div>';
+				}
+			} else if (isset($_GET['success']))
+			{
+				echo'<div>Welcome in Nova !<a href="login_form.php">Connect</a></div>';
+			}
+?>
+
+    <form method="post" action=register_form.php class="buttons1_RegisterForm">
+            <input type="email" class="Register0_RegisterForm" name="email" label="email" id="Register0_RegisterForm" placeholder="your email here..." /></button><br>
+            <input type="password" class="Register1_RegisterForm" name="password" label="Register" id="Register1_RegisterForm" placeholder="type your password..."/></button><br>
+            <input type="password" class="Register2_RegisterForm" name="password_two" label="Register" id="Register2_RegisterForm" placeholder="check your password..."/></button>
+
+            <button type="submit" class="Register_loginEnter" name="RegisterEnter" label="Register" id="RegisterRegister_loginEnter">Register</button>
     </form>
-    
+
+
+
     <div class="arrowBack" onclick="location.href='./index_login.php'">
             <svg version="1.1" id="Calque_1"   x="0px" y="0px"
 	 width="97.411px" height="97.68px" viewBox="0 0 97.411 97.68" enable-background="new 0 0 97.411 97.68" xml:space="preserve">
@@ -61,55 +135,17 @@ require('src/connect.php');
 		c-7.941,0-14.438-6.497-14.438-14.438V16.173c0-7.941,6.497-14.438,14.438-14.438h65.334c7.941,0,14.438,6.497,14.438,14.438
 		v65.334C95.81,89.447,89.313,95.944,81.373,95.944z"/>
 </g>
-            </svg>
-        </div>
-
-            <div class="disclaimer">
-<p class="txt1">Not yet in our Galaxy? <a href="register_form.php">Click here to join!</a></p>
-            </div>
+                </svg>
+    </div>
+    <div class="disclaimer">
+<p class="txt1">Already an account: <a href="./login_form.php">Click here to login!</a></p>
+<p class="txt1"> Having trouble to connect? Reset <a href="./forgot_password.php">password</a></p>
 </div>
+
+</div>
+
 </body>
+<footer>
 
+</footer>
 </html>
-
-<?php
-
-if (isset($_POST['email']))
-    {
-    $newPassword = uniqid();
-    $hashedNewPassword = "aq1".sha1($newPassword."123")."25";
-
-    $sql = "UPDATE user SET password = ? WHERE email = ?";
-    $stmt = $db->prepare($sql);
-    $stmt->execute([$hashedNewPassword, $_POST['email']]);
-    try
-        {    
-    // envoi du mail
-    require "src/Exception.php";
-    require "src/PHPMailer.php";
-    require_once "src/SMTP.php";
-
-    $mail = new PHPMailer(true);
-    //configuration
-
-    $mail->SMTPDebug = SMTP::DEBUG_SERVER; // information de debug
-    $mail->isSMTP();
-    $mail->Host = "sql11.freesqldatabase.com";
-    $mail->Port = 3306;
-    $mail->CharSet = "utf-8";
-    $mail->SMTPSecure = 'ssl';
-    $mail->addAddress($_POST['email']);
-    $mail->setFrom("novaflixbecode@gmail.com");
-    $mail->Subject = "New password - NOVA";
-    $mail->Body = "Hello, here is the new password : .$newPassword.";
-
-    $mail->send();
-    echo 'new password just sent';
-        }
-    catch (Exception $e)
-        {
-    echo "message non envoyé. Erreur: {$mail->ErrorInfo}";
-        }
-    }
-    
-?>
